finishes group index update to use model.js, starts property update. next, finish property forms.

+++ lib/tasks/db.rake
@@ -3,7 +3,7 @@ namespace :db do
 
   desc "Dumps the database to db/APP_NAME.dump"
   task :dump => :environment do
-    cmd = nil
+    cmd = ""
     with_config do |app, host, db, user|
       return "database not found" if db.blank?
       cmd = "pg_dump "
@@ -22,10 +22,22 @@ namespace :db do
   end
 
   desc "Restores the database dump at db/APP_NAME.dump."
-  task :restore => :environment do
-    cmd = nil
+  task :restore, [:arg1] => :environment do |t, args|
+    puts "Args: #{args}"
+    abort("no in file specified to import") if args[:arg1].blank?
+    cmd = ""
     with_config do |app, host, db, user|
-      cmd = "pg_restore --verbose --host #{host} --username #{user} --clean --no-owner --no-acl --dbname #{db} #{Rails.root}/db/#{app}.dump"
+      abort "database not found" if db.blank?
+      cmd = "pg_restore "
+      unless host.blank?
+        cmd += "--host #{host} "
+      end
+      unless user.blank?
+        cmd += "--username #{user} "
+      end
+      cmd += "--clean --no-owner --no-acl --dbname \"#{db}\" \"#{Rails.root}/db/#{args[:arg1]}\""
+      # cmd += app.blank? ? "#db_dump_#{DateTime.now.strftime('%y%m%d_%H%M')}.dump" : "#{app}_#{DateTime.now.strftime('%y%m%d_%H%M')}.dump"
+      # cmd = "pg_restore --verbose --host #{host} --username #{user} --clean --no-owner --no-acl --dbname #{db} #{Rails.root}/db/#{app}.dump"
     end
     Rake::Task["db:drop"].invoke
     Rake::Task["db:create"].invoke
